+++ src/Client/Request/AbstractHtmlRequest.php
@@ -1,0 +1,30 @@
+<?php
+
+namespace DigipolisGent\API\Client\Request;
+
+use DigipolisGent\API\Client\Uri\UriInterface;
+use GuzzleHttp\Psr7\Request;
+use Psr\Http\Message\RequestInterface;
+
+/**
+ * Abstract request requesting HTML response.
+ *
+ * @package Gent\Services\MailingList\Request
+ */
+abstract class AbstractHtmlRequest extends Request implements RequestInterface
+{
+    /**
+     * Constructor.
+     *
+     * @param \DigipolisGent\API\Client\Uri\UriInterface $uri
+     *   The URI for the request object.
+     */
+    public function __construct(UriInterface $uri)
+    {
+        parent::__construct(
+            MethodType::GET,
+            $uri->getUri(),
+            ['Accept' => AcceptType::HTML]
+        );
+    }
+}