+++ src/Client/Request/AcceptType.php
@@ -1,0 +1,30 @@
+<?php
+
+namespace DigipolisGent\API\Client\Request;
+
+/**
+ * The different Accept types.
+ */
+class AcceptType
+{
+    /**
+     * JSON format.
+     *
+     * @var string
+     */
+    const JSON = 'application/json';
+
+    /**
+     * HTML format.
+     *
+     * @var string
+     */
+    const HTML = 'text/html';
+
+    /**
+     * TEXT format.
+     *
+     * @var string
+     */
+    const TEXT = 'text/plain';
+}