+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All Notable changes to `digipolisgent/api-client` package.
 
-## [Unreleased]
+## [1.1.0]
 
 ### Added
 
@@ -11,6 +11,8 @@ All Notable changes to `digipolisgent/api-client` package.
 * Uri component
 * AbstractRequest
 * Allow handlers to support multiple request types
+* Code of Conduct
+* Security guidelines
 
 ## [1.0.0]
 
@@ -24,4 +26,5 @@ This includes:
 * Interfaces to create services in client packages.
 
 [Unreleased]: https://github.com/digipolisgent/php_package_dg-api-client/compare/master...develop
+[1.1.0]: https://github.com/digipolisgent/php_package_dg-api-client/compare/1.0.0...1.1.0
 [1.0.0]: https://github.com/digipolisgent/php_package_dg-api-client/releases/tag/1.0.0